Distributed cloud: Fix Patch Orchestration
The dcmanager failed to create patch strategy for sub-clouds.
This problem is caused by keyword argument name mismatch in the
vim interfaces that used by the dcmanager.
The argument names that have compute reference have been
changed to worker on the vim side while the dcmanager
still uses the old names.

This update changes all compute references to worker in
dcmanager including the data model.

Closes-Bug:1821081

Change-Id: I8cb2dea286e7f91d9b34cc13c57d40de018f3a9f
Signed-off-by: Tao Liu <tao.liu@windriver.com>