Request forwarding and version matching in dcorch proxy
Need to pass in the application to be the forwarder so when
the version doesn't match the request can continue down the
pipeline.  It was 'mostly' working before since the version
matching was registering all get requests as version requests
and just forwarding them, this commit also restricts that
version regex so that is no longer the case.

Change-Id: I887027a043e2a686770d4ece0ae511e00814be61
Closes-Bug: 1849831
Signed-off-by: Tyler Smith <tyler.smith@windriver.com>