Fix a swact issue on system controller
The swact operations failed few times due to dcorch-engine not
being enabled. At the startup, the dcorch-engine starts the
QuotaManager, which tries to open keystone client session
against the keystone-api-proxy. At this time,
the keystone-api-proxy is not yet running.

This dependency is introduced by an incorrect region name used
in the class EndpointCache. The default keystone client session
should be opened against the keystone in RegionOne.
The dcorch-engine enabling process, has a dependency on keystone
which enables keystone first.

This update corrects the region name in EndpointCache.

Change-Id: I1da34ed489c2a7bd6cf43889bb9e173f070b3fa8
Closes-Bug: 1855791
Signed-off-by: Tao Liu <tao.liu@windriver.com>