DC unintentionally re-setup fernet repo on subcloud
Fernet repo is only supposed to be removed and re-setup when the
subcloud goes from managed to unamanged. An unintentional removed
and re-setup is triggered on initial subcloud add when the state
changes from nothing to unmanaged. This can cause timing issues
like removing the fernet repo as the platform-integ-apps is applying.
In this case, resetting the fernet repo fails keystone requests to
get the registry secrets used to deploy platform-integ-apps.
This commit adds a check to only remove and re-setup fernet repo
when the state changes from managed to unmanaged.

Change-Id: Icf21ff7c5bbcfbf34caba28bc83cd806fbb82e19
Closes-Bug: 1855361
Signed-off-by: Jerry Sun <jerry.sun@windriver.com>