Import subprocess from eventlet.green package
The python kubernetes client requires a newer version of eventlet.
However, newer versions of eventlet have an issue with the subprocess
module, which requires subprocess to be imported from eventlet.green
instead of being imported directly. See:
https://github.com/eventlet/eventlet/issues/413

To allow eventlet to be upversioned (from 0.18.4 to 0.24.1), the
subprocess imports are being changed.

Change-Id: I07a6772282a2ad6a65d86c18df1f4a240d19c509
Story: 2006781
Task: 37583
Signed-off-by: Bart Wensley <barton.wensley@windriver.com>