Remote install of sub-cloud controller-0
This update extends the remote sub-cloud deployment to include
the installation of controller-0, which provides a complete ZTP
solution. This is an optional capability that leverages
Redfish Virtual Media Controller(rvmc).

Optional install-value parameters, are added to the dcmanager
subcloud add command, which provides the data required by the
rvmc tool, and update-iso.sh script.

Once install-values are provided, the dcmanager prepares for the
installation and performs the following:
. Downloads an iso image from the url provided, and creates a new
  bootable image based on the install values.
  The new image contains essential info to config a bootstrap ip
  interface that is used to reach the system controller
. Creates a config file for the rvmc tool
. Creates an ansible override file which is used by the install
  playbook

In the next step, the dcmanager runs the install playbook to
install the controller-0 of the subcloud. Once the installation
is completed, the bootstrapping of the sub-cloud would continue.

Story: 2006980
Task: 37715

Depends-On: https://review.opendev.org/#/c/702786/
Change-Id: Id3a1b97adb83a0da51cd54253bdb77e1d729327e
Signed-off-by: Tao Liu <tao.liu@windriver.com>