Restructure subcloud audit to audit subclouds in parallel
Restructuring the dcmanager subcloud audit to audit subclouds
in parallel by creating a greenthread for each subcloud as it
is audited. This reduces the impact on the audit time when
a particular subcloud audit takes a longer amount of time
(e.g. when the subcloud is powered down).

Also updated the EndpointCache class to use the configured
http_connect_timeout (10s) when creating a keystone session.
This reduces the amount of time it takes to timeout a
failed connection to 10s from several minutes.

Finally, added a significant number of unit tests for the
dcmanager subcloud audit.

Change-Id: Ic56470d2f5232c1481730fe4782d27b34089395b
Closes-Bug: 185494
Signed-off-by: Bart Wensley <barton.wensley@windriver.com>