Remove ptp and ntp from dcorch platform sync audit
Remove ptp and ntp from sync audit as it is not required that they are
in sync with the system controller.  This reduces overhead of the audit
as number of subclouds are increased.  It also appears that these
components were not working since change to remove system wide enabled
status for ptp and ntp by https://review.opendev.org/#/c/682121/

Change-Id: Iafcc351a3f001fa259f54184ad023d2a85701289
Closes-Bug: 1857068
Signed-off-by: Gerry Kopec <gerry.kopec@windriver.com>