Distributed Cloud: Large patch upload failed
The third-party modules that used by the proxy service utilize
the tempfile module to create interim files, under the default
temp location.  As a result, a large patch upload failed due to
"No space left on device".

This update sets TMPDIR environment variable to
/scratch/patch-api-proxy-tmpdir in the patch proxy service,
so that the temp files will not use the default location.

Change-Id: Iab4a8fefbab74047b3ec9f318397b21f75c1c5cb
Closes-Bug: 1861329
Signed-off-by: Tao Liu <tao.liu@windriver.com>