Keystone token and client caching
This update is to change the platform SDK driver to cache the keystone
token and client sessions, also to add the fm client to the platform
SDK in order to remove the sdk.py once nova, neutron and cinder
references are removed (as part of story 2006588).

The keystone client/token is cached per-region vs. the client session,
which is cached per-region & per-thread. This is because, simultaneous
access from different greenthreads to the client socket is not allowed.

The initial sync does not use the cached clients as requesting a
new token is required for each region.

In addition, this update modifies the EndpointCache class to load
the auth plugin once.

Test cases:
1. Manage/unmanage subclouds
2. Platform resources and alarm summary sync
3. Platform resources and alarm summary audit
4. Verify the keystone client/token is cached until the token is
   expired
5. Add/delete subclouds
6. Managed subcloud goes offline/online (power off/on)
7. Managed subcloud goes offline/online (delete/add a static route)

Story: 2007267
Task: 38709

Change-Id: I0842a79838ea0f7a6c16f3b1e69ad0eb1357018a
Signed-off-by: Tao Liu <tao.liu@windriver.com>