Adding DistributedCloud subcloud group feature
A new subcloud group feature has been added as a way to
logically group several subclouds.  This would allow a group
of subclouds to be updated in parallel or in serial, or to
limit how many are updated in parallel.

A 'Default' subcloud group is created in all installations,
and all subclouds are automatically members of that
subcloud group unless explicitly updated to reference another.

The Default subcloud group cannot have its name changed, and
other groups cannot use the same name.

A subcloud group that has subclouds in it, cannot be deleted.

Change-Id: Id0f232966c0128ecf6d87f941abcd42ff5a1a5c1
Story: 2007518
Task: 39301
Signed-off-by: albailey <Al.Bailey@windriver.com>