Prevent sync states from bouncing when a subcloud is managed
When a subcloud is first set to managed, the dcorch endpoint
sync status keeps toggling between the two states - "in-sync"
and "out-of-sync", before settling on "in-sync". This is
because during the audit run, different endpoints (sysinv and
identity) are done in separate greenthreads and each
greenthread processes the resources (e.g. users, projects)
in series. Thus, an enhancement is made to reduce the
unnecessary messaging between dcorch and dcmanager, and to
avoid the bouncing sync states.

Change-Id: I98cec42eb5880d95d54182339dab6b5c951d6cfd
Story: 2007267
Task: 38996
Signed-off-by: Jessica Castelino <jessica.castelino@windriver.com>