Avoid synching SSL certificates to subclouds
The SSL certificate on the SystemController will have a CN and SAN
with OAM IP Address and any related DNS Names for OAM IP Address. If
put on the subclouds, even if the certificate was signed by a trusted
CA, when any client connects to the subcloud OAM IP, the certificate
that comes back will not pass certificate validation because the
certificate does not apply to the subcloud's IP Address. For this
reason, a fix is put in place to stop synching SSL certificates
to the subcloud.

Change-Id: I2f5625eeeb30b9fca025b0d0b4a15d926b700d4d
Closes-Bug: 1865643
Signed-off-by: Jessica Castelino <jessica.castelino@windriver.com>