Upgrade strategy support
This feature is based on sw-update-strategy and attempts
to reuse and refactor that code where appropriate so that
more than just patch can be applied using a strategy.

The 'upgrade' strategy type has a new orchestration thread class.

The strategy 'lock' object is shared among all orchestration threads.

Change-Id: Ic1ec7b3665159295005db2133b136f15c62bbddb
Story: 2007403
Task: 39655
Signed-off-by: albailey <Al.Bailey@windriver.com>