Fix bouncing identity sync status on first manage
When a subcloud has been managed for the first time, it's been observed
that its sync status bounces from in-sync to out-of-sync to in-sync.
This is due to identity sync status going out of sync due to identity
roles and admin user requiring two rounds of syncing to fully match
details between system controller and subcloud.  To fix, force roles and
admin user to sync during subcloud initial sync phase.

Change-Id: Ica0c6d8684055e1f1c1c5719b9b0991a75b75bad
Closes-Bug: 1870434
Signed-off-by: Gerry Kopec <gerry.kopec@windriver.com>