Adding a license install step for subcloud upgrade
The license install step queries the license on the
system controller and installs it on the subclouds.

If the system controller does not have a license,
the license install state is skipped and the strategy
proceeds to the next stage.

On successful completion of the 'installing license' stage
the strategy proceeds to the next stage which is the
'importing load' stage.

Change-Id: I61bc144bdf89302b4ba11fd35e82a9ff64d9c517
Story: 2007403
Task: 39780
Signed-off-by: albailey <Al.Bailey@windriver.com>