Use system software version for load audit.
The current audit logic is based on the software version
of the active load as opposed to the software version of
the system. As a result, the subclouds will be out-of-sync
shortly after the first controller has been upgraded.

In this commit, the audit task checks the system software
version which is only updated after both controllers have
been upgraded. This will ensure no premature subcloud
upgrade during upgrade orchestration.

Tests:
  - Modify the system software version of the central
    cloud to simulate a software upgrade and verify that
    all subclouds are out-of-sync after the load audit.
  - Revert the above change and verify that all subclouds
    are back in-sync after the load audit.
  - Modify the system software version of one of the
    subclouds and verify that it is out-of-sync after
    the load audit.

Closes-Bug: 1879396
Change-Id: Ia10d561730f66f63f3e3113d8c3c56270051e7f3
Signed-off-by: Tee Ngo <tee.ngo@windriver.com>