Close any temp files after upload a patch
Add a cleanup routine in the patch-api-proxy to close
any temporary files that opened by third-party libraries,
during a patch upload.

Verified that the temp files are close and /scratch
usage is not increased after a few of large patch
upload.
Verified that the temp files are close and /scratch
usage is not increased after a failed upload.

Change-Id: I68453653fe09d8b1ad80289ab3ff33e476b3acee
Closes-Bug: 1880587
Signed-off-by: Tao Liu <tao.liu@windriver.com>