CLI command to deploy a subcloud
If deployment failed, the user has no option than to delete and
re-add it. If the user was re-adding the subcloud without
re-installing, it would further result in a bootstrap failure.
Thus, to simplify things, a new CLI command is provided to allow
re-deployment. Furthermore, if the user still chooses to delete
the subcloud and re-add it without a re-install, a better error
message is provided asking them to re-install the host.

CLI:
dcmanager subcloud reconfig <id/name> --deploy-config <file>

Test Cases:
1) Successfully add a subcloud with or without deployment option
2) Fail to re-add a subcloud without re-installation after a failed
   deployment
3) Re-deploy with new CLI command after successful and unsuccessful deployment
4) Re-deploy with new CLI command before and after the subcloud is unlocked
5) Test new CLI command by passing wrong parameters

Change-Id: I9fe7e3791e3887160668281048c3c12a7f40c2af
Partial-Bug: 1864756
Signed-off-by: Jessica Castelino <jessica.castelino@windriver.com>