DC upgrade state handler improvements
1) DC upgrade importing-load state should wait for 'imported'

The importing_load state for software upgrade orchestration was
moving to the next state after the load was 'importing' instead
of 'imported'.  This caused the next state to fail since it required
the load to be in the imported state.

2) Use the strategy options object 'alarm-restriction-type' to
determine if an 'upgrade start' should use a force to ignore alarms
that are not management affecting.

3) Add polling stage after upgrade activation to wait for the activity
to complete before transitioning to the next state.

4) Update the unit tests to mock timeout values so that
running the unit tests is quicker.

Change-Id: Ibaf77d2c33b633f915301344c7a7318a59fe3a9f
Story: 2007403
Task: 40040
Signed-off-by: albailey <Al.Bailey@windriver.com>