Add reboot handling to unlock state for DC upgrades
When an unlock is triggered, the host is rebooted.

The maximum amount of time for this state to complete is 1 hour.
 - Reboot time (30 minutes)
 - pre and post reboot time (30 minutes)

The unlock state handler for upgrade orchestration will
now loop and ignore failures for a period of time, while
waiting for the host to become unlocked / enabled.

Change-Id: Ie7ad8dbfecd6ffbe241d098c8677a8e25f400b07
Story: 2007403
Task: 40048
Signed-off-by: albailey <Al.Bailey@windriver.com>