Add thread loop breakout handling for DC upgrade states
If a state machine handler has internal loops, it can check to
see if the main thread is being stopped, and can break out early.

The states performing this check are:
 - importing load (can take up to 30 minutes)
 - starting (can take up to 30 minutes)
 - lock controller (can take up to 10 minutes)
 - unlock controller (can take up to an hour)
 - activating (can take up to 10 minutes)
 - completing (can take up to 10 minutes)

Added a loop-check for 'starting' state to wait for
upgrade to move to 'started'.

Added a loop-check for 'completing' state to wait for
upgrade to be deleted.

Change-Id: I2d4e77d1baf917738ddc96f3cbaa76249ba17fdf
Story: 2007403
Task: 40105
Signed-off-by: albailey <Al.Bailey@windriver.com>