Set up /var/run/dcmanager as dcmanager's default temp files location
dcmanager call the k8s python client to perform a number of
operations. The k8s python client creates temp files under /tmp and
continues use these tmp files for the life-cycle of the processes.

However systemd-tmpfiles-clean.service will run every day to clean up
files in /tmp dir that are older than 10 days. If the k8s client code
is not triggered for more than 10 days (thus its temp files are not
accessed for more than 10 days), these temp files will be removed as
part of the cleanup. Certain dcmanager operations then starts to fail
with an error that the tmp file is no longer there.

This is a known issue of kubernetes python client:
https://github.com/kubernetes-client/python/issues/765

The commit fixes this issue by setting TMPDIR to /var/run/dcmanager
when sm starts dcmanager-manager.

Change-Id: Ib147c2ab26e303032e18da51a506e3768bc471e0
Closes-Bug: 1883599
Signed-off-by: Andy Ning <andy.ning@windriver.com>