Add a 'type' filter for update strategy API calls and DB filtering
The CLI and manager classes can pass an additional 'type' filter so
that upgrade and patch strategy commands are independent of each other.

This allows the patch orchestration CLI to only show and have control
over patch strategies. Horizon will also only show and manage patch
strategies.

Note: there can still only be one strategy in the DB at one time.
When the api filter attribute for the update type is not passed to the
URL, any strategy type can be queried.

The 'create' and 'show' commands explicitly indicate when a mismatched
strategy type is present.

Change-Id: I40058b8cd5d09fa2b700f0a45d10a6b59b74778c
Story: 2007403
Task: 40348
Signed-off-by: albailey <Al.Bailey@windriver.com>