Adding a wait to the end of migrate_data state
The ansible script triggers a reboot and the state needs to
wait for the reboot to complete, before the state transitions
to the next state.

This adds a retry for the upgrade-activate step, and increases
the total max time for that task to be 15 minutes

This also fixes a logging bug where a failure in one subcloud
causes excessive logs while other subclouds are being processed.

Closes-Bug: 1885132
Change-Id: Iabca2ecb0bacff7c5622c90722fcfb938a71a768
Story: 2007403
Task: 40172
Signed-off-by: albailey <Al.Bailey@windriver.com>