Update resources purge logic
The daily dcorch database purge cron job deletes resources that are
no longer used by any orch jobs but as the subcloud_resource table is
set to cascade delete it also removes subcloud resources that are
still in use. This is then flagged by the next dcorch audit and the
subcloud resources are re-created over the next audit cycle(s).

In this commit, only purge resources that are not referenced by any
orch jobs and subclouds.

Closes-Bug: 1887430
Change-Id: I7179a2fba21f296dd51660bb224f5d51aaf531d8
Signed-off-by: Tee Ngo <Tee.Ngo@windriver.com>