Reduce the patch audit interval
Currently the patch audit assumes it is waking up
every 20s so it will wait through 300/20 audit cycles.
Should the subcloud audit take more than 30 seconds,
the patch audit only runs about every 13-14 minutes.

This update stores a timestamp for the last patch audit
and compare with the current time when deciding if an
audit is required. In addition, it triggers the load
audit when a patch audit is forced.

Story: 2007267
Task: 40342

Change-Id: I29e096fd3f2c82f42263f737390b64452bb606c1
Signed-off-by: Tao Liu <tao.liu@windriver.com>