Firmware audit
As part of the dcmanager, there will be a periodic audit which
scans a number of subclouds in parallel and checks whether the
subcloud has all of the “applied” device images that it should
have (based on the labels the images were applied against and
the host labels in the subcloud), and whether all of the “applied”
device images have been installed on the hosts that they should
be. If either of these is not true, then the subcloud
“firmware image sync status” is considered “out of sync”. This
will result in the subcloud as a whole being considered “out
of sync”.
When a device image is “applied” in sysinv, dcmanager
will be notified and will set the “firmware image sync status”
to “unknown” for all subclouds since it does not know at this
point which subcloud(s) the image needs to be created or applied
or updated in. On the next audit, this sync status will be
updated to “in sync” or “out of sync” as applicable.

Change-Id: Ifbbbeaba0bcd2691332e23d288856b8417c9efa8
Story: 2007875
Task: 40420
Signed-off-by: Jessica Castelino <jessica.castelino@windriver.com>