Update dcmanager upgrade orchestration client access
During upgrade orchestration with long running operations, it is
possible that the keystone token may expire and result in
unauthorized http errors.

In dcmanager upgrades orchestration, the client usage is updated to
retrieve the (cached) keystone client to ensure the checks for the
token expiry are up to date.

The unauthorized http exception handling is made explicit for the long
running client operations.

Change-Id: I281ecc51b1a4349ac865f0c4f6ea8244b7cb1fcc
Story: 2007403
Task: 40452
Signed-off-by: John Kung <john.kung@windriver.com>