Add --force option when managing a subcloud
When managing a subcloud, allow a --force option to be
specified. The force option will allow an offline subcloud
to be managed.

This is necessary in cases where the user has (mistakenly)
unmanaged a subcloud that is offline and the subcloud
cannot be recovered while it is unmanaged. One example is
a failure during DC upgrade orchestration that leaves the
subcloud offline. To rerun upgrade orchestration and
recover the subcloud, the subcloud must be managed.

Story: 2007403
Task: 40466
Change-Id: I864de63b40b1065ffb1c6a025b5b221b793fbaf7
Signed-off-by: Tee Ngo <tee.ngo@windriver.com>