Fixes to support DC FPGA Orchestration
When dcmanager-audit is restarted, it ensures that all endpoints
exist for all subclouds. If a subcloud without the endpoint is
encountered during startup, it adds an entry for it without
requiring the subcloud to be re-added. A flag is also added
on a replicated filesystem to avoid these checks from
being executed more than once on the system controller.

Change-Id: I049404c6a3e881623f3ff905a859a94f03bcf8e1
Story: 2007875
Task: 40471
Signed-off-by: Jessica Castelino <jessica.castelino@windriver.com>