Fixes for DC firmware orchestration related to labels and upload
The applied image labels on the system controller should be the labels
used when applying the image on the subcloud.

The stage that removes images now uses all the applied labels
for the image when removing them.

The code that converts image fields for upload was passing an integer
for the revoke_key_id.  It now converts to a string to prevent an
encode exception.

This commit also enables some unit tests for DC firmware
orchestration that were being skipped.

Closes-Bug: 1890355
Closes-Bug: 1890340
Closes-Bug: 1890324
Change-Id: I1a730b01c0bbaeb5f3e1f6ecfb19f649911da453
Signed-off-by: albailey <Al.Bailey@windriver.com>