Trigger firmware audit immediately once subcloud is managed
When a subcloud is managed, patch audit and firmware audit
are triggered. However, the patch audit on completion resets the
force_firmware_audit variable, which prevents the firmware audit
from immediately being triggered and keeps the subcloud out-of-sync
until the next firmware audit cycle (Once in 1 hour). Thus, a fix
is added to address this issue.

Change-Id: Ie1aa2c7b4cb706d32b17f33bffa17d1004721887
Closes-Bug: 1890294
Signed-off-by: Jessica Castelino <jessica.castelino@windriver.com>