Delete subcloud VIM strategies when DC fpga strategy deleted
As part of an fpga upgrade strategy in a subcloud by the VIM,
that strategy needs to be deleted.  Previously it was being
deleted as part of a the final steps of a successful orchestration,
but there can be scenarios where that orchestration is incomplete,
and therefore these VIM strategies need to be handled.

A new DC orchestration cannot be created until the old one is deleted,
so this change ensures fpga vim strategies are deleted if possible
when the DC orchestration strategy is deleted.

A bug in applying vim strategy has been fixed, and the section removed
since a vim auth exception cannot be raised since a new vim client
is constructed per loop.

Unit tests to invoke the delete strategy have been added.

Tox coverage now generates an html report.

Change-Id: I4f3d7bc3550e12c430d1b1c0ce3dffb298fe31b4
Closes-Bug: 1891521
Signed-off-by: albailey <Al.Bailey@windriver.com>