Add a specific counter to handle activation failure
We intend to retry activation 10 times if it fails(in a total of 45
minutes) in the DC orchestration. As the audit and retry are sharing
the same counter, if the activation fails after 10 minutes, the
orchestrator will not re-attempt the activation.

To handle the activation failure, this commit adds a specific counter
to activate retry in case the activate fails after 10 minutes. After
this commit, the orchestrator will wait up to 45 minutes if the
activation is in progress, and the retries will be limited to either
the audit reaches DEFAULT_MAX_QUERIES or itself reaches
MAX_FAILED_RETRIES.

Tested by activating an AIOSX subcloud by orchestration.

Closes-Bug: 1904934
Signed-off-by: Yuxing Jiang <yuxing.jiang@windriver.com>
Change-Id: Ic37c14ee316e02f690b55b082862a85fc54501bc