Add subcloud audit DB table
In order to allow for subcloud audits to be handled by multiple
processes, add a new DB table to track subcloud audit information
including start and finish times for the various audits as well as
whether a special audit has been requested.

On DB upgrade we populate the new table with entries for each
non-deleted subcloud.

The subcloud creation code is modified to also create a corresponding
subcloud_audit entry.

Story: 2007267
Task: 41336
Signed-off-by: Chris Friesen <chris.friesen@windriver.com>
Change-Id: Id2f2b49eb5b8b1f20b1c128bce4811479bcb96b9