Fix bugs with assigning a group for a subcloud
The subcloud CLI and API support passing a group by
name or ID, however the code was only working if an ID was passed.

Furthermore, it could bypass the validator and allow setting zero
as the group id, even though there is no group with an ID of zero.

This submission includes unit tests for updating the group by ID,
name, and with a number of rejection tests.

Change-Id: I53d20a090cfe031c55802104012cea1cb4ff6cb6
Closes-Bug: 1907488
Signed-off-by: albailey <Al.Bailey@windriver.com>