fix up stale subcloud audits
If something happens and an audit is interrupted after it was
started but before it completed, the timestamps will be messed up
in the DB and a new audit won't be triggered because the audit
manager will think that there's an audit already in progress.

Add a new check that looks for subclouds where the audit has been
"in progress" for long enough that we can be pretty sure something
has gone wrong.  Once we find them, reset the audit finish time
so that a new audit will be triggered at the next audit start time.

Story: 2007267
Task: 41336
Signed-off-by: Chris Friesen <chris.friesen@windriver.com>
Change-Id: I390a99ba2bb06915b60fcfb8865096e1ede36cd3