Add a migrate option for rehoming a subcloud
This commit adds a "migrate" option in dcmanager subcloud add API.
With this option, the dcmanager can invoke a rehome_subcloud playbook
to migrate a subcloud under another central cloud. The playbook will
be installed under:
/usr/share/ansible/stx-ansible/playbooks/rehome_subcloud.yml. The
subcloud info will store in dcmanager database.

The keystone password of "smapi" is required during the subcloud
migration, so it is requested from keyring and added into the
subcloud's overrides file.

Tested:
Happy path:
On a system controller, trigger a successful run of the
rehome_subcloud playbook with command:
dcmanager subcloud add --migrate --bootstrap-address
<bootstrap-address> --bootstrap-values bootstrap-values.yml
In the boostrap-values.yml the systemcontroller_gateway_address should
be set under the management subnet.
The subcloud info is added into the dcmanager database. And its
deploy_status is changed in the sequence: pre-rehome-> rehoming->
complete

Unhappy path:
1 On a system controller, trigger a failed run of the rehome_subcloud
playbook with the correct bootstrap values. The subcloud info is also
stored in the dcmanager database. But the deploy_status is changed
into rehome-failed in the end.
2 The systemcontroller_gateway_address is not under the management
subnet. The rehome_subcloud playbook is not triggered, and a notice
indicates the problem is prompted.
3 Create an error during creating the intermediate ca cert in
cert-manager, subcloud's deploy status changes to "rehome-prep-failed"
in the end.

Depends-On: https://review.opendev.org/c/starlingx/ansible-playbooks/+/780101
Story: 2008774
Task: 42150
Change-Id: I758e102d386a63c7ed318bcc91d66e11d3f56299
Signed-off-by: Yuxing Jiang <yuxing.jiang@windriver.com>