Trigger audits on dcmanager-audit startup
If we swact controllers or patch dcmanager-audit, the audit processes
can potentially be killed while there are audits in progress.

Rather than wait for the full timeout, we trigger an "audit fixup"
on dcmanager-audit startup, and also trigger all of the sub-audits
in case we were doing any of them when the audit processes were
killed.

Story: 2007267
Task: 41336
Signed-off-by: Chris Friesen <chris.friesen@windriver.com>
Change-Id: Ic069b6d10cae77613571faaba10869367182c118