dcmanager update availability on first failed audit
In order to improve the response time in the face of failures,
change the dcmanager aduit code to update the availability
(offline/online) on the first failed audit rather than waiting
for a second failure.

This does mean that an intermittent failure will result in
dcmanager treating the subcloud as "offline".

Story: 2007267
Task: 41336
Signed-off-by: Chris Friesen <chris.friesen@windriver.com>
Change-Id: If71d51e82a0147ba38153bf06a6fa8f54d38ae80