trigger subcloud audit when subcloud becomes managed
When we change a subcloud from unmanaged to managed, trigger all of
the various subcloud audits for that subcloud.

Previously we triggered them for *all* subclouds, but now we have
the ability to do it for just one.

Also change it to only trigger the audits if we successfully change
the subcloud to "managed".  No point in triggering them if we
weren't successful or if it was already managed.

Story: 2007267
Task: 41889
Signed-off-by: Chris Friesen <chris.friesen@windriver.com>
Change-Id: I4494c80080dacb59d330162035bd6b595d8026d2