Trigger all sub-audits when subcloud goes from offline to online
When we detect a subcloud going from offline to online, we can't
assume anything about it anymore so let's trigger all the sub-audits
so that on the next audit we'll know whether it's in sync or not.

(I also re-raised a couple of exceptions to get better error
handling.  Shouldn't ever hit this but you never know...)

Closes-Bug: 1916784
Signed-off-by: Chris Friesen <chris.friesen@windriver.com>
Change-Id: Ie252272ff75077e21a9942f136ebfbd95420c175