Handle in-progress subcloud operations after dcmanager-manager restarts
When a subcloud is added it can go through several phases - install,
bootstrap and deploy. If the dcmanager-manager process is restarted
while the subcloud is being added, the DB would never be updated
and subsequent phases are not completed. This commit fixes this
issue by making dcmanager audit for subclouds in transitory states
on startup and update the deploy state to the appropriate failure
state.

Story: 2007267
Task: 42026
Change-Id: I90a57cef89f1cc04a07fa6482ff2c29edf7be965
Signed-off-by: Jessica Castelino <jessica.castelino@windriver.com>